kill some warnings

+++ branches/avidemux_2.4_branch/avidemux/ADM_userInterfaces/ADM_GTK/ADM_dialog/DIA_contrast.cpp
@@ -164,7 +164,7 @@ uint8_t flyContrast::download (void)
   CHECKBOX (checkbuttonV, param.doChromaV);
   
   GtkAdjustment *adj=gtk_range_get_adjustment (GTK_RANGE(WID(hscaleBright)));
-  param.offset= GTK_ADJUSTMENT(adj)->value;
+  param.offset= (int32_t)GTK_ADJUSTMENT(adj)->value;
   
   adj=gtk_range_get_adjustment (GTK_RANGE(WID(hscaleContrast)));
   param.coef=GTK_ADJUSTMENT(adj)->value;
